pyproject.toml
[project]
name = "shortlink-bot"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "python-telegram-bot==22.7",
    "httpx==0.28.1",
    "cryptography==46.0.6",
]
